backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
requests==2.31.0
supabase==2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
